+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added Habit Week Chart Widget
 - Added dynamic Widget previews for launchers
 - Changed Notification Icon to previous icon
+- Fixed settings UI on landscape mode
 
 ## 5.5.53
 - Fixed Notifications not working #229